Upgrade to Flutter V2 plugin API.
