+++ example/android/app/build.gradle
@@ -64,6 +64,7 @@ android {
         release {
             signingConfig signingConfigs.release
             minifyEnabled false
+            shrinkResources false
             proguardFiles "${background_geolocation.projectDir}/proguard-rules.pro"
         }
     }
