+++ example/android/app/src/main/java/com/transistorsoft/flutterbackgroundgeolocationexample/Application.java
@@ -2,14 +2,9 @@ package com.transistorsoft.flutterbackgroundgeolocationexample;
 
 import android.os.StrictMode;
 
-import com.transistorsoft.flutter.backgroundfetch.BackgroundFetchPlugin;
-import com.transistorsoft.flutter.backgroundgeolocation.FLTBackgroundGeolocationPlugin;
-
 import io.flutter.app.FlutterApplication;
-import io.flutter.plugin.common.PluginRegistry;
-import io.flutter.plugins.GeneratedPluginRegistrant;
 
-public class Application  extends FlutterApplication implements PluginRegistry.PluginRegistrantCallback {
+public class Application  extends FlutterApplication {
     @Override
     public void onCreate() {
         // Strict mode.  Should be disabled on RELEASE.
@@ -30,13 +25,5 @@ public class Application  extends FlutterApplication implements PluginRegistry.P
 
         */
         super.onCreate();
-
-        FLTBackgroundGeolocationPlugin.setPluginRegistrant(this);
-        BackgroundFetchPlugin.setPluginRegistrant(this);
-    }
-
-    @Override
-    public void registerWith(PluginRegistry registry) {
-        GeneratedPluginRegistrant.registerWith(registry);
     }
 }
