+++ example/lib/advanced/home_view.dart
@@ -91,6 +91,7 @@ class HomeViewState extends State<HomeView> with TickerProviderStateMixin<HomeVi
 
     bg.TransistorAuthorizationToken token = await bg.TransistorAuthorizationToken.findOrCreate(orgname, username, ENV.TRACKER_HOST);
 
+
     // 2.  Configure the plugin
     bg.BackgroundGeolocation.ready(bg.Config(
         // Convenience option to automatically configure the SDK to post to Transistor Demo server.
