+++ example/lib/main.dart
@@ -97,6 +97,11 @@ void main() {
     // Sanitize old-style registration system that only required username.
     // If we find a valid username but null orgname, reverse them.
     String orgname = prefs.getString("orgname");
+
+
+
+
+
     String username = prefs.getString("username");
     if (orgname == null && username != null) {
       prefs.setString("orgname", username);
