+++ example/pubspec.yaml
@@ -11,7 +11,7 @@ dependencies:
   unicorndial: ^1.1.4
 
   shared_preferences: ^0.5.6
-  background_fetch: ^0.3.2
+  background_fetch: ^0.4.0
   url_launcher: ^5.2.5
 
   # The following adds the Cupertino Icons font to your application.
