+++ pubspec.yaml
@@ -1,18 +1,24 @@
 name: flutter_background_geolocation
-description: The most sophisticated background location-tracking & geofencing module with battery-conscious motion-detection intelligence for iOS and Android.
-version: 1.4.5
-author: Transistor Software <info@transistorsoft.com>
+description: The most sophisticated background location tracking & geofencing module with battery-conscious motion-detection intelligence for iOS and Android.
+version: 1.5.0
 homepage: https://github.com/transistorsoft/flutter_background_geolocation
 
 dependencies:
   flutter:
     sdk: flutter
-  background_fetch: ^0.3.2
+  background_fetch: ^0.4.0
+
+
 flutter:
   plugin:
-    androidPackage: com.transistorsoft.flutter.backgroundgeolocation
-    pluginClass: FLTBackgroundGeolocationPlugin
+    platforms:
+      android:
+        package: com.transistorsoft.flutter.backgroundgeolocation
+        pluginClass: FLTBackgroundGeolocationPlugin
+      ios:
+        pluginClass: FLTBackgroundGeolocationPlugin
 
 environment:
   sdk: ">=2.0.0-dev.50.0 <3.0.0"
-  flutter: ">=1.2.0 <2.0.0"
+  flutter: ">=1.10.0 <2.0.0"
+
